feat(v1-16): Quality 面板拆分、右面板与杂项页面 Design Token 对齐
Part A: Quality 面板拆分
- QualityGatesPanel 967→4 文件 (163+239+192+100 行)
  - qualityGatesTypes.ts: 类型 + 样式常量
  - QualityCheckItems.tsx: 图标 + 状态指示器 + 检查项行
  - QualityRuleList.tsx: 规则列表 + 设置切换
  - QualityGatesPanel.tsx: 薄壳组件 + 重导出
- QualityPanel 575→2 文件 (235+214 行)
  - QualityPanelSections.tsx: JudgeStatusSection + ConstraintsSection
  - QualityPanel.tsx: 有状态主组件

Part B: Diff 模块已对齐 (零硬编码色值，全部 --color-diff-* tokens)

Part C: 杂项页面
- InfoPanel: PanelHeader + --space-section-gap + Text primitive 标签 (301→232 行)
- ZenMode: --zen-content-max-width 720→760px (与 v1-04 一致)
- ZenModeStatus: --color-bg-raised + --shadow-sm 替代 --color-zen-statusbar-bg
- ShortcutsPanel: kbd pill 使用 --color-bg-raised
- AiSettingsSection: FormField 布局替代手写 label+input

Closes #1195
