+++ apps/desktop/renderer/src/features/quality-gates/index.ts
@@ -1,9 +1,10 @@
 /**
  * Quality Gates Panel feature exports
  */
-export { QualityGatesPanel } from "./QualityGatesPanel";
+export { QualityGatesPanel, QualityGatesPanelContent } from "./QualityGatesPanel";
 export type {
   QualityGatesPanelProps,
+  QualityGatesPanelContentProps,
   CheckGroup,
   CheckItem,
   CheckIssue,
