+++ apps/desktop/renderer/src/styles/tokens.css
@@ -198,7 +198,7 @@
   /* Zen mode — extended tokens */
   --color-zen-hover: rgba(255, 255, 255, 0.05);
   --color-zen-statusbar-bg: rgba(0, 0, 0, 0.5);
-  --zen-content-max-width: 720px;
+  --zen-content-max-width: 760px;
   --zen-content-padding-x: 80px;
   --zen-content-padding-y: 120px;
   --zen-title-size: 48px;
